% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{phoneInput}
\alias{phoneInput}
\title{Create a telephone number input control}
\usage{
phoneInput(inputId, label, value = "", width = NULL, placeholder = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display the label for the control, or NULL for no label.}

\item{value}{Initial Value.}

\item{width}{The width of the input, e.g. '400px', or '100%';see \link[shiny]{validateCssUnit}}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the control. Internet Explorer 8 and 9 do not support this option.}
}
\value{
A phone input control that can be added to a UI definition.
}
\description{
Creates a input for telephone numbers, which is validated using the formance JavaScript Library
}
\seealso{
\link[shiny]{textInput}
}

