% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFormancer.R
\name{useFormancer}
\alias{useFormancer}
\title{Include Formance JavaScript Library}
\usage{
useFormancer()
}
\value{
Returns JavaScript files needed to use Formance in Shiny
}
\description{
Adds JavaScript code from Formance library to Shiny application
}

